requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile
#
aiofiles==0.6.0
    # via matrix-nio
aiohttp==3.8.3
    # via
    #   aiohttp-socks
    #   matrix-nio
aiohttp-socks==0.7.1
    # via matrix-nio
aiolog @ git+https://github.com/jaywink/aiolog.git@0db27c59a1694857f58c2bae6874fb25c6c6f262
    # via -r requirements.in
aiosignal==1.2.0
    # via aiohttp
async-timeout==4.0.2
    # via
    #   aiohttp
    #   aiolog
    #   python-socks
atomicwrites==1.4.1
    # via matrix-nio
attrs==22.1.0
    # via
    #   aiohttp
    #   aiohttp-socks
    #   jsonschema
cachetools==4.2.4
    # via matrix-nio
cffi==1.15.1
    # via python-olm
charset-normalizer==2.1.1
    # via aiohttp
commonmark==0.9.1
    # via -r requirements.in
frozenlist==1.3.1
    # via
    #   aiohttp
    #   aiosignal
future==0.18.3
    # via
    #   matrix-nio
    #   python-olm
h11==0.12.0
    # via matrix-nio
h2==4.1.0
    # via matrix-nio
hpack==4.0.0
    # via h2
hyperframe==6.0.1
    # via h2
idna==3.4
    # via yarl
importlib-resources==5.9.0
    # via jsonschema
jsonschema==4.16.0
    # via matrix-nio
logbook==1.5.3
    # via matrix-nio
matrix-nio[e2e]==0.20.0
    # via -r requirements.in
multidict==6.0.2
    # via
    #   aiohttp
    #   yarl
peewee==3.15.3
    # via matrix-nio
pkgutil-resolve-name==1.3.10
    # via jsonschema
psycopg2==2.9.3
    # via -r requirements.in
pycparser==2.21
    # via cffi
pycryptodome==3.15.0
    # via matrix-nio
pyrsistent==0.18.1
    # via jsonschema
python-olm==3.1.3
    # via matrix-nio
python-socks[asyncio]==2.0.3
    # via aiohttp-socks
pyyaml==6.0
    # via -r requirements.in
unpaddedbase64==2.1.0
    # via matrix-nio
yarl==1.8.1
    # via aiohttp
zipp==3.8.1
    # via importlib-resources
